[Add] 메인디테일 fetch 추가 및 드랍다운 컴포넌트화 - 메인디테일 get, post, delet fetch 함수 구현 - 드랍다운 부분 재사용을 위한 컴포넌트화 - 함수 간소화
