+++ src/components/CreateBoard/CreateBoard.js
@@ -1,0 +1,8 @@
+import React from "react";
+import styled from "styled-components";
+
+const CreateBoard = () => {
+  return <></>;
+};
+
+export default CreateBoard;
